app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy==2.0.36
greenlet>=3.0.0
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
pillow==11.0.0
httpx==0.28.1
openai>=1.57.0
chromadb>=0.5.23

# Dev / tests (install with: pip install -r requirements.txt -r requirements-dev.txt)